+++ README.md
@@ -2,7 +2,7 @@
 
 ![axios version](https://img.shields.io/github/package-json/dependency-version/xyzhanjiang/react-bulma-admin/axios) ![bulma version](https://img.shields.io/github/package-json/dependency-version/xyzhanjiang/react-bulma-admin/bulma) ![react version](https://img.shields.io/github/package-json/dependency-version/xyzhanjiang/react-bulma-admin/react) ![react-router version](https://img.shields.io/github/package-json/dependency-version/xyzhanjiang/react-bulma-admin/react-router-dom) ![redux version](https://img.shields.io/github/package-json/dependency-version/xyzhanjiang/react-bulma-admin/@reduxjs/toolkit) [![PRs welcome](https://img.shields.io/badge/PRs-welcome-ff69b4.svg)](https://github.com/xyzhanjiang/react-bulma-admin/pulls)
 
-React 试玩，练手项目/For practising, for my work
+React practising, for my work
 
 * React
 * React-router
@@ -24,7 +24,7 @@ npm install
 npm run dev
 ```
 
-访问 localhost:8080
+Visit localhost:8080
 
 **Build**
 
@@ -34,18 +34,17 @@ npm run build
 
 ## Structure
 
-``` shell
-project-name
+``` diff
+bulma-admin
 |- index.html
 |- package.json
 |- /src
   |- app.js      // the main application component
   |- index.js    // the entry point file for the application
-  |- /actions
+  |- /common
   |- /components
-  |- /containers
   |- /css        // styles
   |- /images
-  |- /reducers   // redux reducers
+  |- /store      // redux store
   |- /views      // page components
 ```
