+++ src/components/modal.js
@@ -5,14 +5,6 @@ export default function Modal(props) {
   React.useEffect(() => {
     // TODO 抖动
     document.documentElement.classList[props.isShown ? 'add' : 'remove']('is-clipped')
-
-    if (props.isShown) {
-      modal.current.offsetWidth
-      modal.current.classList.add('is-in')
-    } else {
-      // TODO 关闭动画效果
-      modal.current.classList.remove('is-in')
-    }
   })
 
   return (
